feat(ffi): Improve enum marshalling (#22)
These improvements come in two forms:
- The first is properly using the bitmask flags to unmarshall a set of
flags (haskell sum type). Additionally we re-create the bitmask when
marshalling back to C
- Separating out dedicated marshalling and unmarshalling functions
(instead of inlining them in the Storable instances). This accomplishes
two goals: consistency, makes the type signatures explicit. This helps
make the connection between the C types, that are marshalling to and
from, explicit.

One theoretical issue for the bitmask words is if an enum type is
encoded as different size words in different structs, then our
marshalling / unmarshalling functions would become a burden (with their
explicit word sizes in haskell). However this did not prove to be a
problem in the current TB client.