Modify Raw/Client.hs
The comments in `tb_client.h` for the `tb_client_t` struct say that
the memory for this struct must remain "pinned." This means the
address of this struct must remain stable throughout the lifetime of
the request.

A `ForeignPtr` represents references managed by a foreign language --
in this case, `tb_client`. The pointer seems to be managed by
`tb_client` as the documentation of `tb_client_deinit` implies that
the library uses the struct to de-allocate internal data it is
managing.

This change is the first in a series to refactor `withClient` to use
the `ForeignPtr` reference. It is also the beginning of a larger
refactor to decompose `withClient` and associated functions so that
the `Raw` module's scope can be limited to calling into the FFI
modules and client functionality can be implemented at a higher level
with safer APIs and structures.

Added:
  - ClientInitError
  - Client
  - initClient
  - initClientEcho

Also updated Internal/FFI/Client.hsc.

Added:
  - TBClientFinalizerCallback
  - makeClientFinalizer