Modify src/Database/TigerBeetle/Client/Sync.hs
Fix the response parsing for result packets.

The Binary instance was not accounting for struct padding in the
result data.

We still need to handle the other operations and clean up the parser
result type.

Still need to parse the batched result responses.  This change only
parses a single result from the response packet.

We also can't seem to verify the account is actually created.

Added:
  - parse response to the callback

Changed:
  - decodeResponse to use the Storable instances

Removed:
  - Binary instance on TBCreateAccountsResult